requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.7.1
cryptography==42.0.7
